// src/trackAssessment/enums/trackAssessment.enum.ts
export enum Status {
  NONE = "", // can be empty to maintain backward compatibility
  COMPLETED = "COMPLETED",
  ABSENT = "ABSENT",
}

export enum trackAssessmentType {
  oral = "oral",
  written = "written",
}

export enum source {
  diksha = "diksha",
  khanAcademy = "khanAcademy",
}

export enum subjectList {
  hindi = "hindi",
  mathematics = "mathematics",
  bengali = "bengali",
  marathi = "marathi",
  english = "english",
  balbharatikannada = "balbharatikannada",
  cocurricular = "cocurricular",
  khelokarusikhu = "khelokarusikhu",
  tamil = "tamil",
  marathishikshaksanhita = "marathishikshaksanhita",
  khelukarushiku = "khelukarushiku",
  playdolearn = "playdolearn",
  telugu = "telugu",
  sindhi = "sindhi",
  gujarati = "gujarati",
  urdu = "urdu",
  khelekaresikhe = "khelekaresikhe",
  kannada = "kannada",
  geography = "geography",
  evs = "evs",
  evspart1 = "evspart1",
  evspart2 = "evspart2",
  scholarshipmarathi = "scholarshipmarathi",
  scholarshipintelligencetest = "scholarshipintelligencetest",
  scholarshipenglish = "scholarshipenglish",
  scholarshipmathematics = "scholarshipmathematics",
  historyandcivics = "historyandcivics",
  sugambharatikannada = "sugambharatikannada",
  science = "science",
  sanskrit = "sanskrit",
  history = "history",
  arabicbidaytularebia = "arabicbidaytularebia",
  defencestudies = "defencestudies",
  kumarbharatikannada = "kumarbharatikannada",
  historyandpoliticalscience = "historyandpoliticalscience",
  selfdevelopment = "selfdevelopment",
  spanish = "spanish",
  gulhaafarsi = "gulhaafarsi",
  secretarialpracticesp = "secretarialpracticesp",
}